% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mixmodfun.R
\name{ModelMixte}
\alias{ModelMixte}
\alias{modelMixte}
\title{Faire une modele mixte Mulitinomial et Gaussien}
\usage{
modelMixte(data, k)
}
\arguments{
\item{data}{data matrix}

\item{k}{number of clusters}
}
\value{
liste avec les proportions(pi), parametres(alpha,mu,cov_list), probabilite a posteriori(tik),BIC,ICL
}
\description{
Faire une modele mixte Mulitinomial et Gaussien
}
